+++ tests/Integration/Auth/CanDirectiveDBTest.php
@@ -23,7 +23,7 @@ final class CanDirectiveDBTest extends DBTestCase
         $this->be($admin);
 
         $user = factory(User::class)->create();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
 
         $this->schema = /** @lang GraphQL */ '
         type Query {
@@ -197,7 +197,7 @@ final class CanDirectiveDBTest extends DBTestCase
     public function testFindUsingNestedInputWithDotNotation(): void
     {
         $user = factory(User::class)->create();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
         $this->be($user);
 
         $this->schema = /** @lang GraphQL */ '
@@ -242,10 +242,10 @@ final class CanDirectiveDBTest extends DBTestCase
         $this->be($admin);
 
         $author = factory(User::class)->create();
-        assert($author instanceof User);
+        $this->assertInstanceOf(User::class, $author);
 
         $post = factory(Post::class)->make();
-        assert($post instanceof Post);
+        $this->assertInstanceOf(Post::class, $post);
         $post->user()->associate($author);
         $post->save();
 
@@ -283,12 +283,12 @@ final class CanDirectiveDBTest extends DBTestCase
         $this->be($admin);
 
         $postA = factory(Post::class)->make();
-        assert($postA instanceof Post);
+        $this->assertInstanceOf(Post::class, $postA);
         $postA->user()->associate($admin);
         $postA->save();
 
         $postB = factory(Post::class)->make();
-        assert($postB instanceof Post);
+        $this->assertInstanceOf(Post::class, $postB);
         $postB->user()->associate($admin);
         $postB->save();
 
@@ -333,7 +333,7 @@ final class CanDirectiveDBTest extends DBTestCase
         $this->be($admin);
 
         $task = factory(Task::class)->create();
-        assert($task instanceof Task);
+        $this->assertInstanceOf(Task::class, $task);
         $task->delete();
 
         $this->schema = /** @lang GraphQL */ '
@@ -373,7 +373,7 @@ final class CanDirectiveDBTest extends DBTestCase
         $this->be($admin);
 
         $user = factory(User::class)->create();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
 
         $this->schema = /** @lang GraphQL */ '
         type Query {
@@ -446,12 +446,12 @@ final class CanDirectiveDBTest extends DBTestCase
         $this->be($admin);
 
         $postA = factory(Post::class)->make();
-        assert($postA instanceof Post);
+        $this->assertInstanceOf(Post::class, $postA);
         $postA->user()->associate($admin);
         $postA->save();
 
         $postB = factory(Post::class)->make();
-        assert($postB instanceof Post);
+        $this->assertInstanceOf(Post::class, $postB);
         $postB->user()->associate($admin);
         $postB->save();
 
@@ -496,7 +496,7 @@ final class CanDirectiveDBTest extends DBTestCase
         $this->be($admin);
 
         $task = factory(Task::class)->create();
-        assert($task instanceof Task);
+        $this->assertInstanceOf(Task::class, $task);
         $task->delete();
 
         $this->schema = /** @lang GraphQL */ '
@@ -579,7 +579,7 @@ final class CanDirectiveDBTest extends DBTestCase
         $company = factory(Company::class)->create();
 
         $user = factory(User::class)->make();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
         $user->company()->associate($company);
         $user->save();