+++ tests/Integration/Execution/DataLoader/RelationBatchLoaderTest.php
@@ -4,6 +4,7 @@ namespace Tests\Integration\Execution\DataLoader;
 
 use Illuminate\Support\Facades\Cache;
 use Nuwave\Lighthouse\Cache\CacheKeyAndTagsGenerator;
+use PHPUnit\Framework\Attributes\DataProvider;
 use Tests\DBTestCase;
 use Tests\Utils\Models\AlternateConnection;
 use Tests\Utils\Models\Comment;
@@ -71,6 +72,7 @@ final class RelationBatchLoaderTest extends DBTestCase
     }
 
     /** @dataProvider batchloadRelationsSetting */
+    #[DataProvider('batchloadRelationsSetting')]
     public function testBatchloadRelations(bool $batchloadRelations, int $expectedQueryCount): void
     {
         $this->schema = /** @lang GraphQL */ '
@@ -438,15 +440,15 @@ final class RelationBatchLoaderTest extends DBTestCase
         ';
 
         $user = factory(User::class)->create();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
 
         $task = factory(Task::class)->make();
-        assert($task instanceof Task);
+        $this->assertInstanceOf(Task::class, $task);
         $task->user()->associate($user);
         $task->save();
 
         $post = factory(Post::class)->make();
-        assert($post instanceof Post);
+        $this->assertInstanceOf(Post::class, $post);
         $post->task()->associate($task);
         $post->user()->associate($user);
         $post->save();
@@ -510,28 +512,28 @@ final class RelationBatchLoaderTest extends DBTestCase
         ';
 
         $user1 = factory(User::class)->create();
-        assert($user1 instanceof User);
+        $this->assertInstanceOf(User::class, $user1);
 
         $post1 = factory(Post::class)->create();
-        assert($post1 instanceof Post);
+        $this->assertInstanceOf(Post::class, $post1);
 
         $comments1 = factory(Comment::class, 3)->make();
         foreach ($comments1 as $comment) {
-            assert($comment instanceof Comment);
+            $this->assertInstanceOf(Comment::class, $comment);
             $comment->user()->associate($user1);
             $comment->post()->associate($post1);
             $comment->save();
         }
 
         $user2 = factory(User::class)->create();
-        assert($user2 instanceof User);
+        $this->assertInstanceOf(User::class, $user2);
 
         $post2 = factory(Post::class)->create();
-        assert($post2 instanceof Post);
+        $this->assertInstanceOf(Post::class, $post2);
 
         $comments2 = factory(Comment::class, 3)->make();
         foreach ($comments2 as $comment) {
-            assert($comment instanceof Comment);
+            $this->assertInstanceOf(Comment::class, $comment);
             $comment->user()->associate($user2);
             $comment->post()->associate($post2);
             $comment->save();