+++ tests/Integration/Execution/MutationExecutor/BelongsToTest.php
@@ -4,6 +4,7 @@ namespace Tests\Integration\Execution\MutationExecutor;
 
 use Illuminate\Database\Events\QueryExecuted;
 use Illuminate\Support\Facades\DB;
+use PHPUnit\Framework\Attributes\DataProvider;
 use Tests\DBTestCase;
 use Tests\Utils\Models\Role;
 use Tests\Utils\Models\Task;
@@ -409,7 +410,7 @@ final class BelongsToTest extends DBTestCase
     public function testCreateAndUpdateBelongsTo(): void
     {
         $user = factory(User::class)->make();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
         $user->name = 'foo';
         $user->save();
 
@@ -449,7 +450,7 @@ final class BelongsToTest extends DBTestCase
     public function testUpsertUsingCreateAndUpdateBelongsTo(): void
     {
         $user = factory(User::class)->make();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
         $user->name = 'foo';
         $user->save();
 
@@ -539,7 +540,7 @@ final class BelongsToTest extends DBTestCase
         ' . self::PLACEHOLDER_QUERY;
 
         $user = factory(User::class)->make();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
         $user->name = 'foo';
         $user->company()->dissociate();
         $user->team()->dissociate();
@@ -604,7 +605,7 @@ final class BelongsToTest extends DBTestCase
     public function testUpsertUsingCreateAndUpdateUsingUpsertBelongsTo(): void
     {
         $user = factory(User::class)->make();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
         $user->name = 'foo';
         $user->save();
 
@@ -650,10 +651,11 @@ final class BelongsToTest extends DBTestCase
     }
 
     /** @dataProvider existingModelMutations */
+    #[DataProvider('existingModelMutations')]
     public function testUpdateAndDisconnectBelongsTo(string $action): void
     {
         $task = factory(Task::class)->create();
-        assert($task instanceof Task);
+        $this->assertInstanceOf(Task::class, $task);
 
         $task->user()->create();
 
@@ -696,12 +698,12 @@ GRAPHQL
     public function testCreateUsingUpsertAndDisconnectBelongsTo(): void
     {
         $user = factory(User::class)->create();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
 
         $task = $user->tasks()->save(
             factory(Task::class)->make(),
         );
-        assert($task instanceof Task);
+        $this->assertInstanceOf(Task::class, $task);
 
         $this->graphQL(/** @lang GraphQL */ '
         mutation {
@@ -742,15 +744,16 @@ GRAPHQL
     }
 
     /** @dataProvider existingModelMutations */
+    #[DataProvider('existingModelMutations')]
     public function testUpdateAndDeleteBelongsTo(string $action): void
     {
         $user = factory(User::class)->create();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
 
         $task = $user->tasks()->save(
             factory(Task::class)->make(),
         );
-        assert($task instanceof Task);
+        $this->assertInstanceOf(Task::class, $task);
 
         $this->graphQL(/** @lang GraphQL */ <<<GRAPHQL
         mutation {
@@ -833,15 +836,16 @@ GRAPHQL
     }
 
     /** @dataProvider existingModelMutations */
+    #[DataProvider('existingModelMutations')]
     public function testDoesNotDeleteOrDisconnectOnFalsyValues(string $action): void
     {
         $user = factory(User::class)->create();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
 
         $task = $user->tasks()->save(
             factory(Task::class)->make(),
         );
-        assert($task instanceof Task);
+        $this->assertInstanceOf(Task::class, $task);
 
         $this->graphQL(/** @lang GraphQL */ <<<GRAPHQL
         mutation {