+++ tests/Integration/Execution/MutationExecutor/HasManyTest.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Integration\Execution\MutationExecutor;
 
+use PHPUnit\Framework\Attributes\DataProvider;
 use Tests\DBTestCase;
 use Tests\Utils\Models\CustomPrimaryKey;
 use Tests\Utils\Models\Role;
@@ -125,10 +126,10 @@ final class HasManyTest extends DBTestCase
     public function testCreateWithConnectHasMany(): void
     {
         $task1 = factory(Task::class)->create();
-        assert($task1 instanceof Task);
+        $this->assertInstanceOf(Task::class, $task1);
 
         $task2 = factory(Task::class)->create();
-        assert($task2 instanceof Task);
+        $this->assertInstanceOf(Task::class, $task2);
 
         $this->graphQL(/** @lang GraphQL */ '
             mutation ($input: CreateUserInput!) {
@@ -322,6 +323,7 @@ final class HasManyTest extends DBTestCase
     }
 
     /** @dataProvider existingModelMutations */
+    #[DataProvider('existingModelMutations')]
     public function testCreateHasMany(string $action): void
     {
         factory(User::class)->create();
@@ -362,10 +364,11 @@ final class HasManyTest extends DBTestCase
     }
 
     /** @dataProvider existingModelMutations */
+    #[DataProvider('existingModelMutations')]
     public function testUpdateHasMany(string $action): void
     {
         $user = factory(User::class)->create();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
 
         $user->tasks()
             ->save(
@@ -409,10 +412,11 @@ final class HasManyTest extends DBTestCase
     }
 
     /** @dataProvider existingModelMutations */
+    #[DataProvider('existingModelMutations')]
     public function testUpsertHasMany(string $action): void
     {
         $user = factory(User::class)->create();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
 
         $user->tasks()
             ->save(
@@ -457,10 +461,11 @@ GRAPHQL
     }
 
     /** @dataProvider existingModelMutations */
+    #[DataProvider('existingModelMutations')]
     public function testDeleteHasMany(string $action): void
     {
         $user = factory(User::class)->create();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
 
         $user->tasks()
             ->save(
@@ -497,16 +502,17 @@ GRAPHQL
     }
 
     /** @dataProvider existingModelMutations */
+    #[DataProvider('existingModelMutations')]
     public function testConnectHasMany(string $action): void
     {
         $user = factory(User::class)->create();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
 
         $task1 = factory(Task::class)->create();
-        assert($task1 instanceof Task);
+        $this->assertInstanceOf(Task::class, $task1);
 
         $task2 = factory(Task::class)->create();
-        assert($task2 instanceof Task);
+        $this->assertInstanceOf(Task::class, $task2);
 
         $actionInputName = ucfirst($action);
 
@@ -555,17 +561,18 @@ GRAPHQL
     }
 
     /** @dataProvider existingModelMutations */
+    #[DataProvider('existingModelMutations')]
     public function testDisconnectHasMany(string $action): void
     {
         $user = factory(User::class)->create();
 
         $taskDisconnect = factory(Task::class)->make();
-        assert($taskDisconnect instanceof Task);
+        $this->assertInstanceOf(Task::class, $taskDisconnect);
         $taskDisconnect->user()->associate($user);
         $taskDisconnect->save();
 
         $taskKeep = factory(Task::class)->make();
-        assert($taskKeep instanceof Task);
+        $this->assertInstanceOf(Task::class, $taskKeep);
         $taskKeep->user()->associate($user);
         $taskKeep->save();
 
@@ -837,10 +844,10 @@ GRAPHQL
     public function testUpdateNestedHasMany(): void
     {
         $user = factory(User::class)->create();
-        assert($user instanceof User);
+        $this->assertInstanceOf(User::class, $user);
 
         $task = factory(Task::class)->create();
-        assert($task instanceof Task);
+        $this->assertInstanceOf(Task::class, $task);
 
         $this->graphQL(/** @lang GraphQL */ '
             mutation ($input: UpdateUserInput!) {