+++ tests/Unit/Schema/Types/Scalars/DateTimeTest.php
@@ -2,12 +2,11 @@
 
 namespace Tests\Unit\Schema\Types\Scalars;
 
-use Nuwave\Lighthouse\Schema\Types\Scalars\DateScalar;
 use Nuwave\Lighthouse\Schema\Types\Scalars\DateTime;
 
 final class DateTimeTest extends DateScalarTestBase
 {
-    protected function scalarInstance(): DateScalar
+    protected function scalarInstance(): DateTime
     {
         return new DateTime();
     }